make resolver thread_func a member function instead of a std::function member variable
There is no point in keeping the resolver function as a member, because it is used
only in a proxy_server constructor, so it can be a local variable of
the proxy_server constructor. In fact there is no point in it being a variable, because
it can be made a function.